FEAT: update error handling in `split_order_into_suborders`
If any order's amount is below the minimum allowed, return an error for that
specific order.

Add unit tests to validate the correct behavior for both single and multiple
ArbitrageOrders.

TODO:

Integrate `welcome_menu` when the bot is sopped and/or paused.

Include parameters `OrderType` and `CurrencyOfInterest` within `welcome_menu`

Author: BugGreen