FIX: Handle minimum order amount error in `run_arbitrage_flow` to prevent infinite loop
Added a check in run_arbitrage_flow to detect when place_sub_orders returns an error with
code "ERROR_BELOW_MIN_TOTAL".

When this error is encountered, the flow logs an error message and exits the loop, preve-
nting an infinite loop.

Provided unit tests to simulate the error case, ensuring that the error is logged and the
flow terminates as expected.

TODO:

Update `MIN_AMOUNT_BINANCE_REQUIREMENTS` and `MIN_WITHDRAWAL_AMOUNT_BINANCE`
with information from the exchange.

Integrate `welcome_menu` when the bot is sopped and/or paused.

Include parameters `OrderType` and `CurrencyOfInterest` within `welcome_menu`

Author: BugGreen