REAFCATOR: Update `split_order_into_suborders` to handle multiple ArbitrageOrder instances
Introduce helper method `_split_order_for_single` to handle single order logic.

Return sub-orders as a list for multiple orders, or a single result for a single
 order.

Added unit tests to verify the functionality of handling one or more orders.

TODO:

Integrate `welcome_menu` when the bot is sopped and/or paused.

Include parameters `OrderType` and `CurrencyOfInterest` within `welcome_menu`

Author: BugGreen