FEAT: `arbitrage_ui` can display the progress of two `ArbitrageOrder` objects
If mode "BOTH_SIDES" is selected, the ui can display the
progress of both orders.

Introduce attribute `order_number` within `ArbitrageOrder` class to
keep track of the number of completed orders.

TODO:

Integrate Websocket for balances and order book

Integrate `welcome_menu` when the bot is stopped and/or paused.

Author: BugGreen